+++ tests/test.bash
@@ -360,7 +360,7 @@ odoo-helper install js-tools;
 
 
 # Install oca/partner-contact addons
-odoo-helper fetch --oca partner-contact;
+odoo-helper fetch --git-single-branch --oca partner-contact;
 
 # Regenerate Ukrainian translations for partner_firstname addons
 odoo-helper tr regenerate --lang uk_UA --file uk_UA partner_firstname;
@@ -397,7 +397,7 @@ odoo-helper-addons-update base
 
 # Fetch OCA account-financial-reporting, which seems to have
 # complicated enough dependencies for this test
-odoo-helper fetch --oca account-financial-reporting
+odoo-helper fetch --git-single-branch --oca account-financial-reporting
 
 # Clone repository explicitly and link it
 (cd repositories && \
@@ -412,7 +412,7 @@ odoo-helper addons update-list
 
 # Generate requirements and fetch them again
 odoo-helper addons generate-requirements > /tmp/odoo-requirements.txt
-odoo-helper fetch --requirements /tmp/odoo-requirements.txt
+odoo-helper fetch --git-single-branch --requirements /tmp/odoo-requirements.txt
 
 # Try to reinstall virtualenv and run server
 odoo-helper install reinstall-venv;
@@ -524,7 +524,7 @@ Fetch OCA/web repo
 ==================
 ${NC}"
 # Fetch oca/web passing only repo url and branch to fetch command
-odoo-helper fetch https://github.com/oca/web --branch 11.0;
+odoo-helper fetch https://github.com/oca/web --branch 11.0 --git-single-branch --git-depth-1;
 
 echo -e "${YELLOWC}
 ============================================
@@ -881,14 +881,81 @@ odoo-helper addons update-list --tdb;
 odoo-helper addons install --tdb --module crm;
 odoo-helper addons test-installed crm;
 
-## Reinstall venv without backup and build for python 3.9.7
-# Python compiling does not work because conflict with bashcov test coverage util
-#odoo-helper install reinstall-venv --no-backup --build-python 3.9.7;
-
-#odoo-helper lsd;  # List databases
+odoo-helper lsd;  # List databases
 
 ## Install addon website via 'odoo-helper install'
-#odoo-helper install website;
+odoo-helper install website;
+
+## Fetch oca/contract
+odoo-helper fetch --github crnd-inc/generic-addons
+
+## Install addons from OCA contract
+odoo-helper addons install --ual --dir ./repositories/crnd-inc/generic-addons;
+
+## Fetch bureaucrat_helpdesk_lite from Odoo market and try to install it
+odoo-helper fetch --odoo-app bureaucrat_helpdesk_lite;
+odoo-helper addons install --ual bureaucrat_helpdesk_lite;
+
+## Print list of installed addons
+odoo-helper addons find-installed;
+
+## Run tests for helpdesk lite
+odoo-helper test generic_request crnd_wsd
+
+# Drop created databases
+odoo-helper db drop odoo15-odoo-test;
+
+echo -e "${YELLOWC}
+=================================
+Install and check Odoo 16.0 (Py3)
+=================================
+${NC}"
+
+cd ../;
+
+# Remove odoo 15
+# this is needed to bypass gitlab.com limitation of disk space for CI jobs
+rm -rf ./odoo-15.0
+
+# Install odoo 16
+odoo-helper install sys-deps -y 16.0;
+
+odoo-install --install-dir odoo-16.0 --odoo-version 16.0 \
+    --http-port 8569 --http-host local-odoo-16 \
+    --db-user odoo16 --db-pass odoo --create-db-user \
+    --build-python-if-needed
+
+cd odoo-16.0;
+
+# Install py-tools and js-tools
+odoo-helper install py-tools;
+odoo-helper install js-tools;
+
+odoo-helper server run --stop-after-init;  # test that it runs
+
+# Show project status
+odoo-helper status;
+odoo-helper server status;
+odoo-helper start;
+odoo-helper ps;
+odoo-helper status;
+odoo-helper server status;
+odoo-helper stop;
+
+# Show complete odoo-helper status
+odoo-helper status  --tools-versions --ci-tools-versions;
+
+# Database management
+odoo-helper db create --tdb --lang en_US;
+
+odoo-helper addons update-list --tdb;
+odoo-helper addons install --tdb --module crm;
+odoo-helper addons test-installed crm;
+
+odoo-helper lsd;  # List databases
+
+## Install addon website via 'odoo-helper install'
+odoo-helper install website;
 
 ## Fetch oca/contract
 #odoo-helper fetch --github crnd-inc/generic-addons
@@ -907,7 +974,8 @@ odoo-helper addons test-installed crm;
 #odoo-helper test generic_request crnd_wsd
 
 # Drop created databases
-odoo-helper db drop odoo15-odoo-test;
+odoo-helper db drop odoo16-odoo-test;
+
 
 echo -e "${YELLOWC}
 =============================================================